+++ CHANGELOG.md
@@ -7,6 +7,9 @@ Warning: Before deploying in production a bot that can generate heavy traffic, p
 All notable changes to Rainbow-Node-SDK will be documented in this file.
 
 ## [1.66.0] - 2020-01-28
+-   Fix when the SDK is already stopped when stop method is called, then return a succeed.
+
+## [1.66.0] - 2020-01-28
 -   Add correlatorData et GlobaleCallId properties in Call type of phone calls : RQRAINB-2773, RQRAINB-2784, RQRAINB-2784, RQRAINB-2789, RQRAINB-2793, RQRAINB-2793, RQRAINB-2799
 -   Fix method ChannelsService::createItem when parameter "type" is setted.
 -   Split Xmmpp error event treatment in 3 possibilities: